# packages/types/src/generated/hTTPError.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HomeOps API
 * HomeOps backend API — household management SaaS (payments, documents, obligations).
 * OpenAPI spec version: 0.1.0
 */
import type { HTTPErrorDetail } from "./hTTPErrorDetail";

export interface HTTPError {
  detail?: HTTPErrorDetail;
  message?: string;
}