# packages/types/src/generated/loginIn.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HomeOps API
 * HomeOps backend API — household management SaaS (payments, documents, obligations).
 * OpenAPI spec version: 0.1.0
 */

export interface LoginIn {
  email: string;
  /**
   * @minLength 1
   * @maxLength 128
   */
  password: string;
}