# packages/types/src/generated/membershipOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HomeOps API
 * HomeOps backend API — household management SaaS (payments, documents, obligations).
 * OpenAPI spec version: 0.1.0
 */

export interface MembershipOut {
  household_id?: string;
  household_name?: string;
  role?: string;
}