# packages/types/src/generated/validationError.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HomeOps API
 * HomeOps backend API — household management SaaS (payments, documents, obligations).
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationErrorDetail } from "./validationErrorDetail";

export interface ValidationError {
  detail?: ValidationErrorDetail;
  message?: string;
}